third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/getComputedStyle-with-before-element-expected.txt

Check getComputedStyle for ::before with various combinations of display and content.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computed.color is 'rgb(0, 128, 0)'
PASS computed.display is 'inline'
PASS computed.color is 'rgb(0, 128, 0)'
PASS computed.display is 'none'
PASS computed.color is 'rgb(0, 128, 0)'
PASS computed.display is 'none'
PASS computed.color is 'rgb(0, 128, 0)'
PASS computed.display is 'inline'
PASS computed.color is 'rgb(0, 128, 0)'
PASS computed.display is 'inline'
PASS computed.color is 'rgb(0, 128, 0)'
PASS computed.display is 'inline'
PASS successfullyParsed is true

TEST COMPLETE
